# ez-client-go/api/ezcards-api-v2-openapi.yaml
openapi: 3.0.3
info:
  title: EZ Cards API Reference V2
  description: B2B Digital Gift Cards Provider - API Reference V2
  version: 2.0.0
  contact:
    name: EZ Group
  license:
    name: Copyright © 2025 EZ Group. All rights reserved.
servers:
  - url: https://api.ezcards.io
    description: EZ Cards API
  - url: https://api.example.com
    description: Example (documentation)
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key assigned to your account for service access.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token to authenticate the API request.
  schemas:
...
    MoneyResponse:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          description: The monetary amount.
        currency:
          type: string
          description: The currency of amount (ISO-4217).
    OrderStatus:
      type: string
      enum: [PROCESSING, COMPLETED, CANCELLED]
      description: |
        PROCESSING - The order is currently being processed.
        COMPLETED - The order has been completed successfully.
        CANCELLED - The order has been cancelled and will not be processed.

    # --- Get Balances ---
...
    PaginationResponse:
      type: object
      required: [total, page, pageSize, totalPage, items]
      properties:
        total:
          type: integer
          description: Total number of items available.
        page:
          type: integer
          description: Current page number.
        pageSize:
          type: integer
          description: Number of items per page.
        totalPage:
          type: integer
          description: Total number of pages available.
        items:
          type: array
          items: {}
          description: List of items on the current page (type varies by endpoint).
...
    OrdersPaginationResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/GetOrdersOrderResponse'
              description: List of orders (Get Orders schema; no payWithCurrency/fxSummary/unitPriceOriginal).

    # --- Create Bulk Order / Create Instant Order / Get Orders ---
    ProductRequest:
      type: object
      required: [sku, quantity]
      properties:
        sku:
          type: string
          description: The Stock Keeping Unit identifier for the product.
        quantity:
          type: integer
          description: The quantity of the product requested in the order.

    # Product in order response (Create Bulk Order & Create Instant Order only – has unitPriceOriginal, FX)
    OrderProductResponse:
      type: object
      required: [sku, quantity, unitPrice, totalPrice, status]
      properties:
        sku:
          type: string
          description: The SKU of the product.
        quantity:
          type: integer
          description: The quantity of the product (always 1 for Instant orders).
        unitPriceOriginal:
          allOf:
            - $ref: '#/components/schemas/MoneyResponse'
          nullable: true
          description: Unit price in the product's native currency before FX. Only in Create Bulk Order / Create Instant Order responses; omitted when same-currency.
        unitPrice:
          $ref: '#/components/schemas/MoneyResponse'
        totalPrice:
          $ref: '#/components/schemas/MoneyResponse'
        status:
          $ref: '#/components/schemas/OrderStatus'

    # Product in order (Get Orders only – doc has no unitPriceOriginal/payWithCurrency/fxSummary)
    OrderProductResponseGetOrders:
      type: object
      required: [sku, quantity, unitPrice, totalPrice, status]
      properties:
        sku:
          type: string
          description: The SKU of the product.
        quantity:
          type: integer
          description: The quantity of the product.
        unitPrice:
          allOf:
            - $ref: '#/components/schemas/MoneyResponse'
          description: The unit price of the product.
        totalPrice:
          allOf:
            - $ref: '#/components/schemas/MoneyResponse'
          description: The total price of the product.
        status:
          $ref: '#/components/schemas/OrderStatus'
    FxEntry:
      type: object
      required: [pair, rate, asOfDate]
      properties:
        pair:
          type: string
          description: Currency pair used (e.g., EUR/USD).
        rate:
          type: number
          description: Applied FX rate for the pair.
        asOfDate:
          type: string
          description: ISO date/time when the rate was effective.

    # Order response for Create Bulk Order & Create Instant Order (has payWithCurrency, fxSummary, product.unitPriceOriginal)
    OrderResponse:
      type: object
      required: [transactionId, clientOrderNumber, grandTotal, createdAt, status, products]
      properties:
        transactionId:
          type: string
          description: The transaction identifier.
        clientOrderNumber:
          type: string
          description: The client's order number.
        payWithCurrency:
          type: string
          description: The currency used to settle/charge the order (Create Bulk/Instant only).
        grandTotal:
          $ref: '#/components/schemas/MoneyResponse'
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the order was created.
        status:
          $ref: '#/components/schemas/OrderStatus'
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProductResponse'
        fxSummary:
          allOf:
            - type: array
              nullable: true
              items:
                $ref: '#/components/schemas/FxEntry'
          description: Summary of FX pairs/rates when cross-currency conversion is applied. Omitted for same-currency orders.

    # Order response for Get Orders only (doc: no payWithCurrency, no fxSummary, product has no unitPriceOriginal)
    GetOrdersOrderResponse:
      type: object
      required: [transactionId, clientOrderNumber, grandTotal, createdAt, status, products]
      properties:
        transactionId:
          type: string
          description: The transaction identifier.
        clientOrderNumber:
          type: string
          description: The client's order number.
        ezOrderNumber:
          type: string
          description: EZ order number. Appears in doc example only; not listed in schema table.
        grandTotal:
          allOf:
            - $ref: '#/components/schemas/MoneyResponse'
          description: The grand total amount and currency.
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the order was created.
        status:
          $ref: '#/components/schemas/OrderStatus'
        products:
          allOf:
            - type: array
              items:
                $ref: '#/components/schemas/OrderProductResponseGetOrders'
          description: List of products in the order.
    CreateBulkOrderRequest:
      type: object
      required: [products]
      properties:
        clientOrderNumber:
          type: string
          description: The order number assigned by the client, used for tracking and reference purposes.
        enableClientOrderNumberDupCheck:
          type: boolean
          default: false
          description: Enable to prevent duplicate clientOrderNumber.
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductRequest'
          description: List of products included in the order.
        payWithCurrency:
          type: string
          description: |
            Settlement currency (ISO-4217). Required only when the product currency is not in the customer's balance currencies.
            If omitted and product currency is in balances, system defaults to product currency.
...
    RequestIdEnvelope:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
          description: Unique identifier assigned to each request for tracking purposes.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: UNSUPPORTED_CURRENCY
        message:
          type: string
          example: payWithCurrency is required or not supported for this account.
        clientOrderNumber:
          type: string
          example: "1234"
paths:
...
  /v2/orders:
    get:
      tags: [Orders]
      summary: Get Orders
      description: Retrieve an order list with pagination support. Use it to check the latest status of specific orders.
      operationId: getOrders
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            description: The number of items to be shown per page.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            description: Page number to retrieve.
        - name: transactionId
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by transaction id for order lookups.
        - name: clientOrderNumber
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by clientOrderNumber for order lookups.
      responses:
        '200':
          description: OK - The list of orders is returned successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RequestIdEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/OrdersPaginationResponse'
        '400':
          description: Bad Request - The request was invalid, often due to incorrect query parameters.
        '500':
          description: Internal Server Error

    post:
      tags: [Orders]
      summary: Create Bulk Order
      description: |
        Create a new bulk order by specifying products and quantities. Operates asynchronously; fulfillment occurs subsequently.
        One API order can only contain products in a single currency.
      operationId: createBulkOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBulkOrderRequest'
            example:
              clientOrderNumber: po-1234
              products:
                - sku: TTR-WI-YWN
                  quantity: 1
                - sku: ELX-BF-85S
                  quantity: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RequestIdEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad Request
        '409':
          description: Conflict (e.g., duplicate clientOrderNumber when duplication check is enabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: CLIENT_ORDER_NUMBER_DUPLICATE
                clientOrderNumber: "1234"
        '422':
          description: Unprocessable Entity (e.g., invalid or missing payWithCurrency when required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: UNSUPPORTED_CURRENCY
                message: payWithCurrency is required or not supported for this account.
        '500':
          description: Internal Server Error
...