# ez-client-go/model_create_bulk_order_request.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreateBulkOrderRequest struct {
	// The order number assigned by the client, used for tracking and reference purposes.
	ClientOrderNumber string `json:"clientOrderNumber,omitempty"`
	// Enable to prevent duplicate clientOrderNumber.
	EnableClientOrderNumberDupCheck bool `json:"enableClientOrderNumberDupCheck,omitempty"`
	// List of products included in the order.
	Products []ProductRequest `json:"products"`
	// Settlement currency (ISO-4217). Required only when the product currency is not in the customer's balance currencies. If omitted and product currency is in balances, system defaults to product currency. 
	PayWithCurrency string `json:"payWithCurrency,omitempty"`
}