# ez-client-go/model_fx_entry.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type FxEntry struct {
	// Currency pair used (e.g., EUR/USD).
	Pair string `json:"pair"`
	// Applied FX rate for the pair.
	Rate float64 `json:"rate"`
	// ISO date/time when the rate was effective.
	AsOfDate string `json:"asOfDate"`
}