# ez-client-go/model_get_orders_order_response.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type GetOrdersOrderResponse struct {
	// The transaction identifier.
	TransactionId string `json:"transactionId"`
	// The client's order number.
	ClientOrderNumber string `json:"clientOrderNumber"`
	// EZ order number. Appears in doc example only; not listed in schema table.
	EzOrderNumber string `json:"ezOrderNumber,omitempty"`
	GrandTotal *MoneyResponse `json:"grandTotal"`
	// Timestamp of when the order was created.
	CreatedAt time.Time `json:"createdAt"`
	Status *OrderStatus `json:"status"`
	// List of products in the order.
	Products []OrderProductResponseGetOrders `json:"products"`
}