# ez-client-go/model_money_response.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type MoneyResponse struct {
	// The monetary amount.
	Amount string `json:"amount"`
	// The currency of amount (ISO-4217).
	Currency string `json:"currency"`
}