# ez-client-go/model_orders_pagination_response.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type OrdersPaginationResponse struct {
	// List of orders (Get Orders schema; no payWithCurrency/fxSummary/unitPriceOriginal).
	Items []GetOrdersOrderResponse `json:"items"`
	// Total number of items available.
	Total int32 `json:"total"`
	// Current page number.
	Page int32 `json:"page"`
	// Number of items per page.
	PageSize int32 `json:"pageSize"`
	// Total number of pages available.
	TotalPage int32 `json:"totalPage"`
}