# ez-client-go/model_product_request.go
/*
 * EZ Cards API Reference V2
 *
 * B2B Digital Gift Cards Provider - API Reference V2
 *
 * API version: 2.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ProductRequest struct {
	// The Stock Keeping Unit identifier for the product.
	Sku string `json:"sku"`
	// The quantity of the product requested in the order.
	Quantity int32 `json:"quantity"`
}